Move into root data ns
