+++ BudgetAnalyser.Engine.XUnit/Ledger/MapperBankBalanceAdjustmentToDto_ToModel_Test.cs
@@ -1,7 +1,6 @@
 ﻿using BudgetAnalyser.Engine.BankAccount;
 using BudgetAnalyser.Engine.Ledger;
 using BudgetAnalyser.Engine.Ledger.Data;
-using BudgetAnalyser.Engine.Ledger.Data.Data2;
 using FluentAssertions;
 using NSubstitute;
