+++ BudgetAnalyser.Engine/Ledger/Data/MapperBankBalanceToDto2.cs
@@ -1,7 +1,7 @@
 ﻿using BudgetAnalyser.Engine.BankAccount;
 using Rees.TangyFruitMapper;
 
-namespace BudgetAnalyser.Engine.Ledger.Data.Data2;
+namespace BudgetAnalyser.Engine.Ledger.Data;
 
 internal class MapperBankBalanceToDto2(IAccountTypeRepository accountTypeRepository) : IDtoMapper<BankBalanceDto, BankBalance>
 {
